Time: 0 ms (100%), Space: 8.5 MB (93.04%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
 
-   double solve(double x, long long n) {
+   double solve(double x, int  n) {
 
         if (n == 0)
             return 1;
